Add pm_init() function, calling which is necessary to use libnetpbm
I've also added an explanatory comment on this with reference to docs.

Note, in order to actually compile the pbm headers and .so, must be visible to gcc.
I.e. in /lib /usr/lib, etc.

When I built and installed from source, they were installed to /usr/local/netpbm/include/ and
/usr/local/netpbm/lib/libnetpbm.so.10.47 respectively. Hence I export $INC and $LIBS on the
shell, and Makefile has been designed to pick them up.

For now, I think this is a reasonable compromise depending on the user, if they are in the
standard path then gcc will pick them up fine, if not the user can export the variable (or
hack the Makefile, if they so desire). Or softlink.

TODO: Add an explanatory note, probably in a proper readme rather than code.