Add README for netpbm_bitflip
The netpbm_copy code was relatively easy to describe comments, but in order to understand how the a netpbm image represents image data
(and thus how we can manipulate it) some discussion is necessary. This is too much to easily fit in code comments, hence this README.

Note, it is still a work in progress, I'm committing this now to ensure initial work isn't lost.